requirements.txt
asgiref==0.14.0
autobahn==0.16.0
channels==0.17.2
daphne==0.15.0
Django==1.9.9
django-appconf==1.0.2
django-bower==5.1.0
django-compressor==2.1
django-crispy-forms==1.6.0
django-parsley==0.6
django-registration==2.1.2
futures==3.0.5
rcssmin==1.0.6
rjsmin==1.0.12
git+git://github.com/S-Intelligent-Technologies/django-builder.git
six==1.10.0
Twisted==16.4.0
txaio==2.5.1
zope.interface==4.3.0
asgi_redis
PyMySQL==0.7.8
python-memcached
asgi_ipc
django-anymail[mailgun]==0.5
django-allauth